migrator: log directly to persistent storage
The initramfs logs default to being written to tmpfs at /run. The
migrator hook also writes messages to this location, which is later
copied to persistent storage by resin-init-flasher if present. However,
any failure before resin-init-flasher, such as failing to copy
installation binaries to memory, will result in the file not being
written to persistent storage for debugging.

Write the log file to persistent storage directly in the migrator hook
in case of failure before resin-init-flasher, but also write to tmpfs
and copy to the new boot partition filesystem in the flasher script.

This ensures the migration logfile is always available on persistent
storage no matter where a failure occurs.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>