resin-init-flasher: use configuration file to specify post-install action
Traditionally after programming the flasher image would shutdown the
device to allow for the removal of the external device and booting from
the internal storage.

As more use case appeared, like programming from the initramfs when the
flasher image runs on the same target device it will program, the post
install action needs to support rebooting instead as physical access to
the device to power on cannot be expected. This is the case for remote
migration, AMI VM installs and also secure boot installs.

The current code just checks for the existence of the `shutdown` command
which is not present on the initramfs to decide whether to shutdown or
reboot.

This commit improves that logic by defaulting to shutdown but allowing the
configuration file to specify a reboot instead.

Change-type: minor
Signed-off-by: Alex Gonzalez <alexg@balena.io>