initrscripts: migrate: use configuration file to specify post-install action
Use the flasher configuration file to specify the post-install action.

This is to shutdown (the default) when the flasher runs from an external
media, or to reboot when running from the same media as is the case for
the migrator tool.

Also, assert that a locked device can only install from memory.

Change-type: minor
Signed-off-by: Alex Gonzalez <alexg@balena.io>